requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyMySQL
requests
gunicorn
python-dotenv
openai
